Replace the queue() function with a static Queue class
This makes the Queue class autoloadable and fixes cases
were a consumer wants to set a queue driver via a bootstrap
file loaded from the autoload/files section of the composer.json.
The bootstrap file might be loaded before the functions.php and
the queue() function is then not available.